switch to f64. -O1 by default
test262: 9.58% | 🧪 48642 | 🤠 4658 | ❌ 108 | 💀 16 | 🧩 188 | 💥 16164 | 📝 27508

+++ compiler/codeGen.js
@@ -1011,7 +1011,7 @@ export default program => {
   depth = [];
   currentFuncIndex = Object.keys(importedFuncs).length;
 
-  globalThis.valtype = 'i32';
+  globalThis.valtype = 'f64';
 
   const valtypeOpt = process.argv.find(x => x.startsWith('-valtype='));
   if (valtypeOpt) valtype = valtypeOpt.split('=')[1];
